magento/inventory#2728: Moved result interface under Api folder
